Interim.  Completed refactoring of Memory Hierarch.
Buffer hierarchy is not done. Many Errors in Test. Main compiles clean,
no warnings.
